backend/api/requirements.txt
fastapi==0.75.1
pydantic==1.9.0
uvicorn==0.17.6
mysql-connector-python==8.0.28